+++ system/modules/isotope/IsotopeProduct.php
@@ -428,6 +428,9 @@ class IsotopeProduct extends Controller
 								$this->arrCache[$strKey] = $this->Database->execute("SELECT page_id FROM tl_iso_product_categories WHERE pid=" . ($this->pid ? $this->pid : $this->id) . " ORDER BY sorting")->fetchEach('page_id');
 								break;
 
+							case 'cssID':
+								$this->arrCache[$strKey] = deserialize($this->arrData[$strKey], true);
+
 							default:
 								return isset($this->arrData[$strKey]) ? deserialize($this->arrData[$strKey]) : null;
 						}