+++ system/modules/isotope/languages/de/tl_iso_products.php
@@ -52,6 +52,8 @@ $GLOBALS['TL_LANG']['tl_iso_products']['shipping_exempt'][0] = 'Vom Versand auss
 $GLOBALS['TL_LANG']['tl_iso_products']['shipping_exempt'][1] = 'Klicken Sie hier wenn dieses Produkt nicht versendet wird (wie z.B. herunterladbare Produkte.';
 $GLOBALS['TL_LANG']['tl_iso_products']['tax_class'][0] = 'Steuerklasse';
 $GLOBALS['TL_LANG']['tl_iso_products']['images'][0] = 'Bilder';
+$GLOBALS['TL_LANG']['tl_iso_products']['cssID'][0] = 'CSS-ID/Klasse';
+$GLOBALS['TL_LANG']['tl_iso_products']['cssID'][1] = 'Hier können Sie eine ID und beliebig viele Klassen eingeben.';
 $GLOBALS['TL_LANG']['tl_iso_products']['published'][0] = 'Produkt veröffentlichen';
 $GLOBALS['TL_LANG']['tl_iso_products']['published'][1] = 'Klicken Sie hier wenn das Produkt auf der Webseite angezeigt werden soll.';
 $GLOBALS['TL_LANG']['tl_iso_products']['start'][0] = 'Startdatum';
@@ -128,6 +130,7 @@ $GLOBALS['TL_LANG']['tl_iso_products']['inventory_legend'] = 'Inventar-Einstellu
 $GLOBALS['TL_LANG']['tl_iso_products']['shipping_legend'] = 'Versand-Einstellungen';
 $GLOBALS['TL_LANG']['tl_iso_products']['options_legend'] = 'Produktoptionen-Einstellungen';
 $GLOBALS['TL_LANG']['tl_iso_products']['media_legend'] = 'Dateiverwaltung';
+$GLOBALS['TL_LANG']['tl_iso_products']['expert_legend'] = 'Experten-Einstellungen';
 $GLOBALS['TL_LANG']['tl_iso_products']['publish_legend'] = 'Veröffentlichung';
 $GLOBALS['TL_LANG']['tl_iso_products']['variant_legend'] = 'Produktvarianten-Einrichtung';